Added get_task
